refactor(product-workflow): split reborn service types
