+++ crates/ironclaw_product_workflow/src/reborn_services/types.rs
@@ -1,0 +1,108 @@
+use ironclaw_host_api::ThreadId;
+use ironclaw_product_adapters::{ProductOutboundEnvelope, ProjectionCursor};
+use ironclaw_threads::{SessionThreadRecord, SummaryArtifact, ThreadMessageRecord};
+use ironclaw_turns::{
+    AcceptedMessageRef, CancelRunResponse, EventCursor, ResumeTurnResponse, TurnRunId, TurnStatus,
+};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RebornCreateThreadResponse {
+    pub thread: SessionThreadRecord,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(tag = "outcome", rename_all = "snake_case")]
+pub enum RebornSubmitTurnResponse {
+    Submitted {
+        thread_id: ThreadId,
+        accepted_message_ref: AcceptedMessageRef,
+        turn_id: String,
+        run_id: TurnRunId,
+        status: TurnStatus,
+        resolved_run_profile_id: String,
+        resolved_run_profile_version: u64,
+        event_cursor: EventCursor,
+    },
+    DeferredBusy {
+        thread_id: ThreadId,
+        accepted_message_ref: AcceptedMessageRef,
+        active_run_id: TurnRunId,
+        status: TurnStatus,
+        event_cursor: EventCursor,
+    },
+    AlreadySubmitted {
+        thread_id: ThreadId,
+        accepted_message_ref: AcceptedMessageRef,
+        run_id: TurnRunId,
+        status: TurnStatus,
+        event_cursor: EventCursor,
+    },
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RebornTimelineRequest {
+    pub thread_id: String,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RebornTimelineResponse {
+    pub thread: SessionThreadRecord,
+    pub messages: Vec<ThreadMessageRecord>,
+    pub summary_artifacts: Vec<SummaryArtifact>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RebornStreamEventsRequest {
+    pub thread_id: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub after_cursor: Option<ProjectionCursor>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RebornStreamEventsResponse {
+    pub events: Vec<ProductOutboundEnvelope>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RebornCancelRunResponse {
+    pub run_id: TurnRunId,
+    pub status: TurnStatus,
+    pub event_cursor: EventCursor,
+    pub already_terminal: bool,
+}
+
+impl From<CancelRunResponse> for RebornCancelRunResponse {
+    fn from(value: CancelRunResponse) -> Self {
+        Self {
+            run_id: value.run_id,
+            status: value.status,
+            event_cursor: value.event_cursor,
+            already_terminal: value.already_terminal,
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RebornResumeGateResponse {
+    pub run_id: TurnRunId,
+    pub status: TurnStatus,
+    pub event_cursor: EventCursor,
+}
+
+impl From<ResumeTurnResponse> for RebornResumeGateResponse {
+    fn from(value: ResumeTurnResponse) -> Self {
+        Self {
+            run_id: value.run_id,
+            status: value.status,
+            event_cursor: value.event_cursor,
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(tag = "outcome", rename_all = "snake_case")]
+pub enum RebornResolveGateResponse {
+    Resumed(RebornResumeGateResponse),
+    Cancelled(RebornCancelRunResponse),
+}
